update sheet.py to follow new template, update configs
